cleanup old code
