+++ autocomplete/__init__.py
@@ -2,6 +2,6 @@
 Allow HTMXAutoComplete to be imported from the module directly
 """
 
-from .autocomplete import Autocomplete, HTMXAutoComplete, register, urls
+from .autocomplete import Autocomplete, register, urls
 from .shortcuts import ModelAutocomplete
 from .widgets import AutocompleteWidget
